Moved meta details to another file and cleaned up generator a bit
