+++ games/__meta__.yaml
@@ -1,0 +1,23 @@
+name: Player{player}
+description: Archipelago Async Mystery Filler Settings
+requires:
+  version: 0.3.2
+  plando: bosses, items, text, connections
+
+# Define your game here!
+game:
+  A Link to the Past: 150
+  Factorio: 20
+  Minecraft: 50
+  Slay the Spire: 20
+  Risk of Rain 2: 30
+  Subnautica: 5
+  Ocarina of Time: 40
+  Timespinner: 50
+  Secret of Evermore: 10
+  Super Metroid: 30
+  Rogue Legacy: 50
+  Super Mario 64: 20
+  VVVVVV: 20
+  Meritous: 10
+  SMZ3: 30
